requirements.txt
numba
numpy
torch
tqdm
more-itertools
tiktoken
triton>=2.0.0;platform_machine=="x86_64" and sys_platform=="linux" or sys_platform=="linux2"
discord.py[voice]
openai-whisper
pydub
numpy
soundfile
discord-ext-audiorec